Homeview, canvas, stickerbar updated
