+++ client/components/Canvas.js
@@ -3,12 +3,28 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 
 class Canvas extends Component {
-  constructor() {
-    super()
+  constructor(props) {
+    super(props)
 
     this.state = {
       canvas: []
     }
+  }
+  componentDidMount() {
+    let canvas = new fabric.Canvas('my-canvas')
+  }
+
+  addToCanvas(sticker) {
+    console.log('clicked', sticker)
+
+    fabric.Image.fromURL(sticker.imgURL, img => {
+      img.scale(0.2)
+      img.set({left: 100, top: 100})
+      canvas
+        .add(img)
+        .renderAll()
+        .setActiveObject(img)
+    })
   }
   // componentDidMount(){
   //   const canvas = new fabric.Canvas('my-canvas')
@@ -23,6 +39,7 @@ class Canvas extends Component {
   // }
 
   render() {
+    console.log('canvas props=', this.props.fabricCanvas)
     return (
       <div>
         <h1> My Canvas</h1>
